Bug fix: output_interval as double instead of int
Former-commit-id: 805a26259e491b236944262bd7b69ab814ff13e3 [formerly 49998257997c2c750db380318eedd1cddbf2805f]
Former-commit-id: 1cc8401ef5a84c91ec5721a787fc11afe644f871

+++ src/integrator.c
@@ -407,7 +407,7 @@ IN_FILE ErrorStatus euler(
 
     bool is_output = (output_param->method != OUTPUT_METHOD_DISABLED);
     int *__restrict output_count_ptr = &(output_param->output_count_);
-    const int output_interval = output_param->output_interval;
+    const double output_interval = output_param->output_interval;
     double next_output_time = output_interval;
 
     double *__restrict t_ptr = &(simulation_status->t);
@@ -583,7 +583,7 @@ IN_FILE ErrorStatus euler_cromer(
 
     bool is_output = (output_param->method != OUTPUT_METHOD_DISABLED);
     int *__restrict output_count_ptr = &(output_param->output_count_);
-    const int output_interval = output_param->output_interval;
+    const double output_interval = output_param->output_interval;
     double next_output_time = output_interval;
 
     double *__restrict t_ptr = &(simulation_status->t);
@@ -758,7 +758,7 @@ IN_FILE ErrorStatus rk4(
 
     bool is_output = (output_param->method != OUTPUT_METHOD_DISABLED);
     int *__restrict output_count_ptr = &(output_param->output_count_);
-    const int output_interval = output_param->output_interval;
+    const double output_interval = output_param->output_interval;
     double next_output_time = output_interval;
 
     double *__restrict t_ptr = &(simulation_status->t);
@@ -1023,7 +1023,7 @@ IN_FILE ErrorStatus leapfrog(
 
     bool is_output = (output_param->method != OUTPUT_METHOD_DISABLED);
     int *__restrict output_count_ptr = &(output_param->output_count_);
-    const int output_interval = output_param->output_interval;
+    const double output_interval = output_param->output_interval;
     double next_output_time = output_interval;
 
     double *__restrict t_ptr = &(simulation_status->t);
